requirements.txt
feedparser==6.0.11
requests==2.32.3
trafilatura==1.12.0
groq==0.8.0
google-generativeai==0.7.2
gspread==6.1.2
google-auth==2.29.0
pyyaml==6.0.1
python-dateutil==2.9.0
